{권원중} - hrm, department 유효성검사

+++ src/main/java/com/goodee/groupware/controller/DepartmentController.java
@@ -31,6 +31,13 @@ public class DepartmentController {
 		Map<String,Object> resultMap = departmentService.getDepartmentList();
 		String memberId = (String) session.getAttribute("loginMember");
 		model.addAttribute("memberId", memberId);
+		// 부서관리는 팀장급만 들어올 수 있습니다.
+		String Level = (String) session.getAttribute("memberLevel");
+		// 앞에 숫자만 사용하여 비교하도록 int값으로 파싱함
+		int memberLevel = Integer.parseInt(Level.substring(0, 1));
+		if(memberLevel < 3) {
+			return "redirect:/group/home";
+		}
 		// Model에 addAttribute를 사용하여 view에 값을 보낸다.
 		// 부서 리스트
 		model.addAttribute("departmentList", resultMap.get("department"));
@@ -46,7 +53,7 @@ public class DepartmentController {
 		public String addDepartment(Department department) {
 			// 유효성검사
 			if(department.getDepartmentId().equals("") || department.getDepartmentId() == null) {
-				
+				return "redirect:/group/department/departmentList";
 			}
 			// 현재부서의 가장 높은 번호 가져오기
 			int MaxDepartmentNo = departmentService.getMaxDepartmentNo(department);
@@ -64,9 +71,13 @@ public class DepartmentController {
 		}
 	// 부서 이동
 		@PostMapping("/group/department/updateDepartment")
-		public String updateDepartment( @RequestParam(name="littleDepartment", defaultValue="0") int departmentNo,
+		public String updateDepartment( @RequestParam(name="littleDepartment", defaultValue="-2") int departmentNo,
 										@RequestParam(name="memberId") String memberId){
 			log.debug("DepartmentCotroller.@RequestParam()-->" + departmentNo + memberId);
+			// 유효성 검사
+			if(memberId.equals("") || memberId == null) {
+				return "redirect:/group/department/departmentList";
+			}
 			Member member = new Member();
 			member.setMemberId(memberId);
 			member.setDepartmentNo(departmentNo);
@@ -78,7 +89,9 @@ public class DepartmentController {
 	// 부서 삭제
 		@PostMapping("/group/department/deleteDepartment")
 		public String deleteDepartments(Department department) {
-			
+			if(department.getDepartmentNo() <= 0 || department.getDepartmentNo() == 100 || department.getDepartmentNo() == 200) {
+				return "redirect:/group/department/departmentList";
+			}
 		    int deletedCount = departmentService.deleteDepartments(department);
 		    log.debug("DepartmentCotroller.deletedCount()-->" + deletedCount);
 		    return "redirect:/group/department/departmentList";
